refactor: move useResponses hook to the Operation feature folder
