+++ packages/api-reference/src/components/OperationPath.vue
