+++ packages/api-reference/src/embeds/ExampleRequest/ExampleRequest.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import { useServerStore } from '#legacy'
+import OperationPath from '@/components/OperationPath.vue'
 import { ExampleRequest } from '@/features/ExampleRequest'
-import OperationPath from '@/features/Operation/components/OperationPath.vue'
 import { TestRequestButton } from '@/features/TestRequestButton'
 import type { Spec, TransformedOperation } from '@scalar/types/legacy'
 import { toRef } from 'vue'
